extract ApplyOffsetRewriter from ApplyFilterTransposeRule
