+++ src/frontend/src/optimizer/rule/mod.rs
@@ -97,6 +97,9 @@ pub use avoid_exchange_share_rule::*;
 mod min_max_on_index_rule;
 pub use min_max_on_index_rule::*;
 
+mod apply_offset_rewriter;
+use apply_offset_rewriter::ApplyOffsetRewriter;
+
 #[macro_export]
 macro_rules! for_all_rules {
     ($macro:ident) => {
